Fixing project errors
